requirements.txt
pytest~=7.2.0
requests~=2.28.1
testit-pytest
testit-api-client
selenium-wire
selenium==4.7.2
self~=2020.12.3
urllib3~=1.26.12
datetime~=4.9
allure-pytest
jsonschema
docker
behave~=1.2.6
pillow
